wip: attempt at solving recent error
current tf apply error: `creating CloudFront Distribution: InvalidViewerCertificate: The certificate that is attached to your distribution doesn't cover the alternate domain name (CNAME) that you're trying to add`

This is just an idea for a fix. Not sure if it helps since I don't have access to run tf apply yet.